requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==24.3.0
certifi==2024.12.14
click==8.1.8
discord.py==2.4.0
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
multidict==6.1.0
ollama==0.4.6
propcache==0.2.1
pydantic==2.10.5
pydantic_core==2.27.2
PyYAML==6.0.2
sniffio==1.3.1
typing_extensions==4.12.2
watchdog==6.0.0
yarl==1.18.3
